Set UDS horizontal scaling phase to auto
Use the register default settings for the scaling phase instead
of manually calculating them.

The default setting seems to calculate the phase correctly,
which the manual settings seem to shift the output image to the left
my an amount proportional to the scaling ratio.  i.e x16 scaling
will shift the output by 16 pixels.